refactor: reorganized package and override toString in TokenDto.kt
