+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/controller/auth/AuthController.kt
@@ -1,8 +1,8 @@
-package br.com.igorrodrigues.farmcontrol.controller
+package br.com.igorrodrigues.farmcontrol.controller.auth
 
 import br.com.igorrodrigues.farmcontrol.domain.model.User
-import br.com.igorrodrigues.farmcontrol.domain.usecase.CreateUserUseCase
-import br.com.igorrodrigues.farmcontrol.domain.usecase.UserDto
+import br.com.igorrodrigues.farmcontrol.domain.usecase.user.CreateUserUseCase
+import br.com.igorrodrigues.farmcontrol.domain.usecase.user.UserDto
 import br.com.igorrodrigues.farmcontrol.infrastructure.security.TokenDto
 import br.com.igorrodrigues.farmcontrol.infrastructure.security.TokenService
 import org.springframework.http.HttpStatus.CREATED
