+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/domain/usecase/user/CreateUserUseCase.kt
@@ -1,4 +1,4 @@
-package br.com.igorrodrigues.farmcontrol.domain.usecase
+package br.com.igorrodrigues.farmcontrol.domain.usecase.user
 
 import br.com.igorrodrigues.farmcontrol.domain.model.AllUser
 import br.com.igorrodrigues.farmcontrol.domain.model.User
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Service
 
 @Service
 class CreateUserUseCase(private val allUser: AllUser,
-                        private val tenantService: TenantService) {
+                        private val tenantService: TenantService
+) {
 
     fun create(userDto: UserDto): User {
         if (allUser.userAlreadyExist(userDto.email)) throw UserAlreadyExistentException()
