+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/domain/usecase/user/TenantService.kt
@@ -1,4 +1,4 @@
-package br.com.igorrodrigues.farmcontrol.domain.usecase
+package br.com.igorrodrigues.farmcontrol.domain.usecase.user
 
 interface TenantService {
     fun initDatabase(schema: String)
