+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/domain/usecase/user/UserDto.kt
@@ -1,4 +1,4 @@
-package br.com.igorrodrigues.farmcontrol.domain.usecase
+package br.com.igorrodrigues.farmcontrol.domain.usecase.user
 
 data class UserDto(
         val email: String = "",
