+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/infrastructure/persistence/TenantServiceAdapter.kt
@@ -1,6 +1,6 @@
 package br.com.igorrodrigues.farmcontrol.infrastructure.persistence
 
-import br.com.igorrodrigues.farmcontrol.domain.usecase.TenantService
+import br.com.igorrodrigues.farmcontrol.domain.usecase.user.TenantService
 import org.flywaydb.core.Flyway
 import org.springframework.stereotype.Component
 import javax.sql.DataSource
