+++ src/main/kotlin/br/com/igorrodrigues/farmcontrol/infrastructure/security/TokenDto.kt
@@ -3,4 +3,6 @@ package br.com.igorrodrigues.farmcontrol.infrastructure.security
 data class TokenDto(
         val value: String = "",
         val type: String = "Bearer",
-)
+) {
+    override fun toString() = "$type $value"
+}
