+++ src/test/kotlin/br/com/igorrodrigues/farmcontrol/controller/auth/AuthControllerTest.kt
@@ -1,8 +1,9 @@
-package br.com.igorrodrigues.farmcontrol.controller
+package br.com.igorrodrigues.farmcontrol.controller.auth
 
+import br.com.igorrodrigues.farmcontrol.controller.ErrorDetail
 import br.com.igorrodrigues.farmcontrol.domain.model.User
-import br.com.igorrodrigues.farmcontrol.domain.usecase.CreateUserUseCase
-import br.com.igorrodrigues.farmcontrol.domain.usecase.UserDto
+import br.com.igorrodrigues.farmcontrol.domain.usecase.user.CreateUserUseCase
+import br.com.igorrodrigues.farmcontrol.domain.usecase.user.UserDto
 import br.com.igorrodrigues.farmcontrol.infrastructure.security.TokenDto
 import br.com.igorrodrigues.farmcontrol.infrastructure.security.TokenService
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
